Just Another Commit

+++ MicroWars/Tesla.h
@@ -10,16 +10,16 @@ namespace microwars
 		float tesla_pos_x, tesla_pos_y;
 		float tesla_radius;
 		int tesla_health;
-		int tesla_x_factor;
+		float tesla_x_factor;
 
 		public:
 		
-		Tesla(float x, float y, float radius, int x_factor);
+		Tesla(float x, float y, float radius, float x_factor);
 		float return_tesla_pos(char option);
 		int check_unit_vicinity(float x, float y);
 		void change_health();
 		int return_health();
-		int return_x_factor();
+		float return_x_factor();
 
 		~Tesla();
 	};
